futility: update: add new quirk 'no_check_platform'
Some devices may have flashed firmware with different platform name in
their early stage (especially in the first build of leading devices),
so we do want to provide an explicit way (not just --force) to skip
checking platform name.

The change CL:2059621 does not help because the loaded system
firmware looks good.

This is implemented as a quirk so we can enable it using a CBFS quirk
file, making it easier to be deployed by auto update.

BRANCH=None
BUG=None
TEST=make runtests

Change-Id: I888d5848921d31c9b7cba1b96c42d38fda71927e
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/2573999
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>