crossystem: add a hwid override mechanism from chromeos-config
Devices which do not use chromeos firmware (e.g., reven) do not have a
GBB section nor firmware to expose this data via ACPI/device-tree.
However, a hwid is still required for auto-update, UMA, etc.

CL:3118810 adds a new property to the schema for /:hwid-override.
This CL makes it so when that property is set, ACPI/device-tree is not
probed for the HWID, and it is instead set via cros_config.

BUG=b:195007267
BRANCH=none
TEST=build image for reven with a hwid override
     # crossystem hwid
     REVEN-ANAE A6A-A7I

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I3c1255b257f914b02dfb8ba53dd275892df1a526
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/3119046
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>