vboot/ui: Show error box on internal boot failure
CL:2917623 changes the behavior when failing to boot from internal disk
in developer UI. However, it causes the device to potentially enter a
reboot loop (b/197216832).

Instead of either triggering recovery or returning an error from the UI
loop, change the behavior to showing an error message in a dialog box
and staying in the UI. This is similar to the behavior when booting from
an invalid external disk, where an error screen will be shown to the
user.

Add VB2_UI_ERROR_INTERNAL_BOOT_FAILED to vb2_ui_error enum.

BUG=b:197216832, b:197911901
TEST=make runtests
BRANCH=none

Cq-Depend: chromium:3123160
Change-Id: Ic5ace55fc8b93ba0836e5722b7c5011fd490c35e
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/3124395
Reviewed-by: Julius Werner <jwerner@chromium.org>