futility/update: Search RO CBFS files in "COREBOOT", not "RO_SECTION"
The "RO_SECTION" FMAP area is a parent area that contains multiple
nested areas, including the "COREBOOT" area containing the actual RO
CBFS. On some platforms the "COREBOOT" area may be the first area in
"RO_SECTION", and therefore the start offsets would be equal and
`cbfstool -r RO_SECTION` would accidentally do the right thing -- but
this is not true for all platforms and may break elsewhere.

I don't know what this code does or how to test it but I assume fixing
this issue is a good thing.

BRANCH=None
BUG=None
TEST=None

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Ica281af75299e06c36e2f0437d2b1fcc6c0731e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/3299058
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>