OWNERS: Remove `*` approval
Adding `*` to OWNERS makes Rubber Stamper not work for cherry-picks, so
we transition to something else. While coreboot and depthcharge
responsibilities are so distributed that it probably still makes sense
to allow the whole firmware team to approve (see CL:3687245 and
CL:3687246), vboot is not as platform-specific and it seems that we
mostly have a tighter group of contributors. Looking at the recent
commits, it seems that pretty much every patch was approved by one of
the existing explicitly-mentioned owners anyway, so I think it makes
sense to reduce ownership to that set. (Note that the scripts/
subdirectory already has its own additional OWNERS file managed by the
infra team, so we don't necessarily need to cover that.)

BRANCH=none
BUG=none
TEST=none

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Ib902f4dcf654f00703da8fff831c6a2b9f7e9e0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/3687195
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>