gscvd: presume GBB flags are zero when hashing the RO space contents
It is still being debated who is supposed to make sure that the GBB
flags are set to zero before the root of trust validation is granted
to the AP firmware image, but as of today the approach is that the GBB
flags must be zero at AP RO validation time.

The problem is that when AP RO space signature is created GBB flags
can be set to a non-zero value.

With this patch when AP RO areas contents is hashed, in case GBB flags
are included in one of the ranges, the flags are not read from the
flash, and substituted with zero.

During validation the real flags value is used. A unit test is added
to verify various futility gscvd GBB related situations, the blobs for
the unit test were extracted from a Nivviks firmware image.

BRANCH=none
BUG=b:245799496, b:253540670
TEST='./tests/futility/test_gscvd.sh' and 'make runfutiltests' succeed

Change-Id: I2f047b990cf71ea24d191fc690da08e25ebb10cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/3958581
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4659288
Reviewed-by: Yu-An Wang <wyuang@google.com>
Commit-Queue: Yu-An Wang <wyuang@google.com>
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>
Auto-Submit: LIU KUAN-FU <iamjeffliu@google.com>