make_dev_ssd.sh: Add logs of make_dev_ssd.sh
Sometimes we want to know if someone disable rootfs verification
on DUT, but we have no evidence in logs.
This CL modify make_dev_ssd.sh, for logging the operations and
the parameters passed to the script in /var/log/messages when
the device is changing itself.
It can help us checking:
1. The operartions done by executing make_dev_sdd.sh
2. Whether someone disable rootfs verification.
3. The time when the rootfs verification was disabled.

BUG=b:140709784
TEST=manual test on TOMATO DUT

Change-Id: I8140e13b0e74a9b2a084954c5153e528045d7c7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4022414
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4660090
Commit-Queue: Yu-An Wang <wyuang@google.com>
Reviewed-by: Yu-An Wang <wyuang@google.com>
Auto-Submit: LIU KUAN-FU <iamjeffliu@google.com>
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>