sign_official_build: Make reven check more explicit
Add a new get_is_reven() function which uses get_board_from_lsb_release()
to determine the board name. This replaces a glob check against the
KEY_DIR with an exact string match to prevent future accidents (e.g. if
a key set named "Revenator" was added), and also checking against the
board name makes local testing easier (since the test devkeys can be
used without copying to a new "Reven" directory).

Also removed a todo about getting "is reven?" from a config in the
signing repo. The linked bug is closed, and the TODO is 16 months
old now so it's probably fine as-is.

BRANCH=none
BUG=b:256621033
TEST=platform/vboot_reference/scripts/image_signing/sign_official_build.sh
TEST=  recovery build/images/reven/latest/chromiumos_image.bin
TEST=  platform/vboot_reference/tests/devkeys
TEST=  build/images/reven/latest/chromiumos_image.bin.signed

Change-Id: I3586b607447b0f55a4a253d9a4d9c7212889c7d6
Signed-off-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4206907
Reviewed-by: Mike Frysinger <vapier@chromium.org>