lib/flashrom_drv.c: Use correct programmer name for programmer init
Fix the call to `flashrom_programmer_init()` in `flashrom_get_wp()`. The
programmer name and parameters should be passed separately, e.g:

> flashrom_programmer_init("dummy", "emulate=VARIABLE_SIZE,...")

However a string containing both the name and parameters was incorrectly
used instead, e.g:

> flashrom_programmer_init("dummy:emulate=VARIABLE_SIZE,...", "emulate=VARIABLE_SIZE,...")

This was not noticed before because it did not cause issues if there
were no parameters, e.g. if the programmer was just "host" or "internal"

BUG=b:238694831,b:260531154
TEST=cros_run_unit_tests --board octopus --packages vboot_reference

Change-Id: I4774a82258acbb8be2b430351ccef203e26e6b27
Signed-off-by: Nikolai Artemiev <nartemiev@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4290837
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>