tpm2_lite: Treat NV_UNINITIALIZED error as BADINDEX for TPM 2.0
This patch ports the equivalent of CB:20299 to the vboot version of the
TPM library. This should fix a problem that was observed where a created
but not written FWMP space gets a device stuck in recovery mode.

BRANCH=none
BUG=b:272310645
TEST=none

Change-Id: If0ef6fdda879b45de0444cdaeb423d01ce109e2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4327074
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4660169
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>
Auto-Submit: LIU KUAN-FU <iamjeffliu@google.com>
Reviewed-by: Yu-An Wang <wyuang@google.com>
Commit-Queue: Yu-An Wang <wyuang@google.com>