vbutil_kernel: Move kernel's EFI boot stub into bootloader section
This patch modifies vbutil_kernel to treat EFI boot stubs differently
than what was first introduced in CL:3900403. Instead of just leaving
the kernel image unchanged in that case, it will still pick apart the
vmlinuz image to extract the protected mode kernel and move it to the
front of the kernel body section. However, it will then also preserve
the EFI boot stub and instead make it take the place of the "bootloader"
section in the back of the partition, provided that no other explicit
file has been provided with the --bootloader flag. This means that these
images remain compatible with the traditional (depthcharge) method of
booting x86 kernels, but also gives other bootloaders that require the
EFI stub (e.g. crdyboot) a chance to put it back into place.

Also make the --bootstub flag optional because it's not really needed
anymore nowadays (not passing it will just create a zero-length
bootloader section). And deduplicate some code between KernelSize() and
PickApartVmlinuz().

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Ifdbf29511dabc91c235608648221952789e7a95f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4548417
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>