sign_uefi: Remove exception catching
The behavior of allowing sbsign to fail with a warning has existed since
the original CL adding UEFI signing: crrev.com/c/995175. The warning
gets "upgraded" to a fatal error later, when calling sbverify, but
there's no good reason not to fail immediately if sbsign fails, so
remove the try block.

Also remove the try block when calling sbverify so that a failure
results in a normal python stack trace.

BRANCH=none
BUG=b:286234394
TEST=make runtests
TEST=run sign_official_build.sh on a local build

Change-Id: I8b56d2602c53b3eb0ea8a1d6f42fd07d71121635
Signed-off-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4759949
Reviewed-by: Mike Frysinger <vapier@chromium.org>