futility/cmd_show: Support --publickey FW_VBLOCK
We are relying on the 'futility vbutil_firmware' to verify firmware and
extract the kernel subkey from the firmware. However, due to some
technical reasons 'vbutil_firmware' doesn't work with firmware built
with VBOOT_CBFS_INTEGRATION, and there's no plan to support that.

Instead, we'd like to exclusively use 'futility verify' for verifying
firmware and kernel blobs. For firmware verification there is no problem
because everything we need is in the firmware image. However, for kernel
verification we'll need the public kernel subkey stored in the firmware
preamble inside the vblock.

To support that, make the --publickey recognize firmware vblock file
type.

BUG=b:307158019
TEST=make runfutiltests
BRANCH=none

Change-Id: Ibdd1179ec6492581e753d17c248fc3f36bc79c88
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4977847
Reviewed-by: Julius Werner <jwerner@chromium.org>