sign_uefi: Support signing via pkcs11
In sign_uefi.py, if the private key is a pkcs11 URI then add "--engine
pkcs11" to the args sbsign is called with.

In sign_official_build.py, make use of the new KEYCFG_UEFI_* env vars to
set the key paths. This allows the private key to be switched from the
local path to a pkcs11 URI. (Note: kernel_subkey.vbpubk is used in more
places than just UEFI signing, so left that out of this migration for
now.)

BRANCH=None
BUG=b:286234394
TEST=make runtests
TEST=run sign_official_build.sh on a local build
TEST=Run signing_docker's full.sh, in combo with crrev.com/i/6630383

Change-Id: I21202bc01352fc721636ad1cabf5b6b796a08aaf
Signed-off-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5002498
Reviewed-by: Allen Webb <allenwebb@google.com>