futility/cmd_show: Fix output for firmware body signature
The condition 'if (state || (sign_key && good_sig))' near the end of
the show_fw_preamble_buf() is about keyblock signature. However the
output printed in that if block is about body signature inside the
preamble. Since the information about the keyblock signature is already
printed in show_keyblock(), we don't need to print anything at the end
of the function. Fix the problem and add a few test cases for that.

Also fix the prefix of a parseable output message "data_key::invalid".

The bios_brya_mp_invalid_vblock_b.bin file is created from
bios_brya_mp.bin by the following commands:

 futility vbutil_keyblock --pack keyblock.bin \
   --datapubkey tests/devkeys/firmware_data_key.vbpubk --flags 23
 futility sign --keyblock keyblock.bin \
   --infile bios_brya_mp.bin --outfile bios_brya_mp.bin.new
 cbfstool bios_brya_mp.bin.new read -r VBLOCK_B -f vblock_b.bin
 cbfstool bios_brya_mp.bin write -r VBLOCK_B -f vblock_b.bin

BUG=none
TEST=make runfutiltests
BRANCH=none

Change-Id: I9f0b2135b422933e6aa710fcdf6d816c7fad0936
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5007248
Reviewed-by: Julius Werner <jwerner@chromium.org>