tests/futility/test_show_and_verify: Add test cases for invalid data key
Add two test cases for an invalid firmware vblock (fw_pre) with an
invalid data key. Modify the test script to also check the stdout even
if 'futility show' is expected to fail.

Note that '--type fw_pre' needs to be passed for the futility command to
fail. Otherwise the invalid vblock file will be recognized as a keyblock
file.

BUG=none
TEST=make runfutiltests
BRANCH=none

Change-Id: Ie0cab87f6cf3bc150c20eda7785d6663b751a6ee
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5034514
Reviewed-by: Julius Werner <jwerner@chromium.org>