host/lib/pkcs11: Remove superfluous 'nss' directory from include paths
Some distros use "nss" for the directory name containing the header
files and others use "nss3", which breaks the compilation of these files
since the includes use "nss". However, the build system already looks up
the path of the header files and adds it to the searchable paths, so the
explicit use of the nss directory in the includes is not needed. Thus,
just remove it.

Build-tested with Arch Linux and Fedora.

Change-Id: Ic34b29c9a78f90e579b9d6aa12ffd7219aeefa45
Signed-off-by: Felix Singer <felixsinger@posteo.net>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5079884
Reviewed-by: Ching-Kang Yen <chingkang@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@google.com>
Commit-Queue: Patrick Georgi <pgeorgi@google.com>
Reviewed-by: Julius Werner <jwerner@chromium.org>