make_keyblock: Add support for omitting extension
In order to sign with Cloud KMS, vbutil_keyblock supports a
`--signprivate` in the pkcs11 uri format (i.e. `--signprivate
remote:<p11_lib>:<slot_id>:<key_label>`, but in order to support that
flow we need to not hardcode the `.vbprivk` extension. Add an arg that
supports omitting that extension.

BUG=b:318522770
BRANCH=None
TEST=None

Change-Id: I0c32cbf5bb8c47f5e971b5002df604366e3ba0f5
Signed-off-by: Benjamin Shai <bshai@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5182236
Reviewed-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Rachael Newitt <renewitt@google.com>