avb: Add callback for public key verification
In order to establish secure link between platform and software it is
necessary to embed public part of the signing key within the persistent
storage (in our case flash memory). During boot & verification process,
this key is compared by AVB library to the one used to sign vbmeta
partition.

BUG=b:309426555
TEST=With consecutive CLs applied build FW and boot android
BRANCH=main

This CL is based on crrev.com/c/5650817

Change-Id: Ia2dd88bdbfc745f509922369a7c7b3f808c2e48e
Signed-off-by: Jan Dabros <dabros@google.com>
Signed-off-by: Grzegorz Bernacki <bernacki@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/6063630
Tested-by: Konrad Adamczyk <konrada@google.com>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>