sign_uefi_unittest: Refactor in preparation for more tests
Split the test suite into two. One for actual signing tests that require
various test files to be created, and one for utility functions.

In the signing test suite, move the key and EFI file creation into the
setUp function so it can be reused. Split the successful signer test
into two tests, and change the tests to all make use of the files
created in setUp.

The coverage from the tests is unchanged, this is just a refactor to
make it easier to add additional tests.

BRANCH=None
BUG=b:331773545
TEST=make runtests

Change-Id: I5e463335e112aa27fec6a7b851887ded6966c27f
Signed-off-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5405600
Reviewed-by: Mike Frysinger <vapier@chromium.org>