DEDEDE-ONLY: HACK: Record extra debugging info for vboot corruption
This patch adds a function that will check for secdata_kernel
corruption, and record several hints about the system state in the
recovery subcode field when it finds some. This is intended to help us
track down the source of secdata_kernel corruption that was observed on
Metaknight and Cret.

BRANCH=dedede
BUG=b:331989856
TEST=Built for Dedede, booted and manually triggered on a Trogdor

Change-Id: Ibbeef35a4e3c098b9afceb2530ed33415b3c05a2
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5499389
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>