futility: Try to load ecrw versions regardless of image type
CL:5550200 added a new field 'is_ec' to the firmware_image struct.
However, that struct is supposed to hold only basic data for
firmware-like images.

To allow removing the field from the struct, firmware parsing needs to
be independent of the image type (AP or EC). Therefore, change the
parsing logic so that ecrw versions will be loaded whenever the
corresponding FMAP sections FW_MAIN_* are present.

BUG=b:340132454
TEST=make runfutiltests -j
BRANCH=none

Change-Id: Ic649ccf89ed5ca4804d96c30447f711695c3e9d4
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5561607
Reviewed-by: Hung-Te Lin <hungte@chromium.org>