futility/load_fmap: Erase remaining bytes if file smaller than area
The `futility load_fmap` subcommand allows the passed file to be smaller
than the corresponding FMAP area size. That's convenient for some use
cases (such as EC's KEY_RO), because the caller doesn't need to have an
extra step of enlarging the file to match the area size.

However, the current implementation is to keep the remaining data in the
area unmodified, which is often not the intention of the caller.
Therefore, change the behavior by erasing the remaining bytes to 0xff.

BUG=none
TEST=make runfutiltests
BRANCH=none

Change-Id: Ib6575023801d995b7b25110bb411ec74d9a240f0
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5768173
Reviewed-by: Julius Werner <jwerner@chromium.org>