futility: updater: Provide default DUT properties for emulation
When --emulate is specified, all the dut_get_property() calls would
fail. The option --sys_props exists to allow overriding the DUT
properties. However, it's bothersome to specify it every time we want to
run the updater with emulation.

Therefore, provide default DUT properties for emulation. The active
firmware slot is set to A. The TPM firmware version is set to 0x10001
(to pass TPM key compatibility check). The platform version is set to 0.
Both hardware and software write protect states are assumed to be
disabled for convenience.

Some of the --sys_props arguments in test_update.sh are simplified or
omitted.

BUG=none
TEST=make runfutiltests -j
TEST=futility update --emulate image.bin -i new.bin --mode=factory
BRANCH=none

Change-Id: I25202ad05a5c2e671b416c42951be8cdafee5689
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5807011
Reviewed-by: Hung-Te Lin <hungte@chromium.org>