signing: miniOS signing in docker.
For some reason, lsblk reports null partition types in docker, even in
privileged
(https://forums.docker.com/t/why-does-lsblk-report-partition-as-null-from-within-a-docker-container/139393).
Prefer cgpt for checking the partition type, which appears to work, and
fallback on lsblk (because the legacy signers only have that).

This commit also removed the check for the private key on disk because
the key lives in Cloud KMS, adds a newline to an info statement that
was getting concatenated to the previous log message, and unrolls the
loop over all loop devices (opting to just explicitly check the two
minios loop devices).

BUG=b:363466852
BRANCH=None
TEST=manual

Change-Id: I044c626dfac5ae7115b9f6b7e65e8303595544fb
Signed-off-by: Benjamin Shai <bshai@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5833916
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: George Engelbrecht <engeg@google.com>