futility: updater: Process custom label as standard models
One advantage of switching to signer_config.csv and getting rid of
setvars.sh is we finally can know what custom label devices are
supported.

In `setvars.sh` or the pre-Unified-Build single package format, we have
to scan the 'keyset/' folder to figure out what custom label tags may be
available. But when using signer_config.csv, all the custom label names
are already included in the signer information.

As a result, we can finally remove the hacky way of patching custom
label devices. One downside is we can no longer support custom label in
non-Unified-Build systems, but all active Chromebooks are now Unified
Build so that is not a problem. And for development using simple archive
without `signer_config.csv`, there is no need to support custom label
because we usually don't have signed keys.

This is also the first step to remove the custom label logic from the
firmware updater - preparation to let crosid handle custom label
matching for us. This also implies `--signature_id` is now equivalent to
`--model` and can be deprecated in the follow up changes.

BUG=b:251040363
TEST=make runfutiltests -j
BRANCH=none

Change-Id: Ib4c62ba4d01b276472575625ea19397fc21b1589
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5872452
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>