tlcl: Increase TPM buffer size
This change increases the TPM message buffer size from 512 to 1024 in
order to be able to fit new nvmem entry in single read command. Prior
to this change the TlclRead(0x3fff0a) would fail, because the space
size was too big (820 bytes) to fit into response buffer. With this
change the nvmem content can be read safely.

This change is required for CL:5947750.

BUG=b:359340876 b:354045389
TEST=Verify that nvmem contents are read correctly.
BRANCH=firmware-android-15949.B

Change-Id: Id227acc5a40d47ec81c2e2ddc83503763e0c88f3
Signed-off-by: Bartłomiej Grzesik <bgrzesik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5987271
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Grzegorz Bernacki <bernacki@google.com>