crossystem: Change cros_debug to rely on mainfw_type, not devsw_boot
cros_debug?1 is true either when the kernel command line indicates that
we're running a developer image, or when devsw_boot indicates that the
developer switch is on. It is generally used to enable insecure
debugging features that we do not want to be available in secure mode.

Unfortunately, devsw_boot is not the best way to test for this. It
reflects the raw state of the developer switch and returns true
regardless of boot mode (e.g. also in recovery mode). This means that
cros_debug?1 cannot be relied upon to enable debug features in recovery
images because their initramfs environment is supposed to remain secure
even if the developer switch is on.

mainfw_type better reflects what we want here, since it will always be
`recovery` in recovery mode and only return `developer` when booting
from the developer screen. This patch changes cros_debug to test this
instead so it can be reliably used even in recovery contexts.

BRANCH=none
BUG=b:382540412
TEST=Booted CoachZ in secure, developer, recovery and dev+rec modes,
confirmed that cros_debug?1 always reflects the expected state.

Change-Id: I5bd4ca2da081a7ed125a002a72d99f6ee4444715
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/6073625
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Owners-Override: Jae Hoon Kim <kimjae@chromium.org>
(cherry picked from commit c57a588f8029ac83675349e75b539ecdc4bcd280)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/6081934