Makefile: Remove $(shell) invocations from CFLAGS
CFLAGS is a recursive variable, meaning it needs to be re-evaluated
every time it is referenced. By adding a `$(shell)` into CFLAGS we force
`make` to invoke the shell commands every single time. This results in
a lot of unnecessary pkg-config calls.

I straced the process to confirm the fix:

Before:
```
$ grep pkg-config /tmp/coreboot.strace| wc -l
102973
```

After:
```
$ grep pkg-config /tmp/coreboot.fast.strace| wc -l
3937
```

This shaves off ~30 - 60 seconds on my machine when building brya.

BUG=b:392102378
BRANCH=none
TEST=emerge-brya coreboot

Change-Id: Ib23f1821e64f491e1a7929a8d6812c23f6ec85ad
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/6198687
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Julius Werner <jwerner@chromium.org>