sign_gsc_firmware.sh: sing ECDSA only image for NT
The NT ROM_EXT gets confused if Ti50 binary includes some elements of
the PQ signature, even if the SPX key is not included in the owners
configuration.

With the resent update in the build scripts an ECDSA only binary is
included in the tarball, use it for nightly signing instead of the
previously available one.

BRANCH=none
BUG=b:395968753
TEST=verified that locally produced image signed by a Cloud KMS ECDSA
     key boots fine on an NT chip.

Change-Id: Ica147c39bf108eb371446c44e32c17edcaa52205
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/6370072
Reviewed-by: Benjamin Shai <bshai@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Commit-Queue: Vadim Bendebury <vbendeb@google.com>