futility/updater: Add warning for full firmware update without WP
This commit introduces a warning message to the `updater.c` file. When
write protection is disabled, and a full firmware update is about to
proceed, a warning will now be displayed to the user. This message
clarifies that the entire SPI flash, including RO (Read-Only) sections,
will be overwritten.

Additionally, this commit renames the `optarg` variable to `optval` in
`handle_flash_argument` to prevent conflicting with `unistd.h`.

BUG=none
BRANCH=none
TEST=futility update -i zero.bin --mode=autoupdate
TEST=and see the warning

Change-Id: Idf544e02f9d3a53a42049b017178458484ccd6c6
Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/6567851
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>
Tested-by: Yu-Ping Wu <yupingso@chromium.org>