host/flashrom_drv: refactor flashrom initialization
This change refactors the flashrom initialization code. Most of the
functions in flashrom_drv.c had the same copy of it, so in this change
it was moved to a separate function `flashrom_setup`, which handles all
of the relevant errors in a single place. Labels `err_init` and
`err_probe` have been removed, as they are no longer required.

BRANCH=main
BUG=None
TEST=make USE_FLASHROM=1 runtests

Change-Id: I90d70d53ebc54f717467ff0c4f9ff4a22008c2f9
Signed-off-by: Ihor Lukianov <lis05st@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/7044447
Reviewed-by: Jakub "Kuba" Czapiga <czapiga@google.com>
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>
Commit-Queue: Jakub "Kuba" Czapiga <czapiga@google.com>