futility: Update --model help text
The `--model` argument is not only for debugging and testing purposes.
For example, in Android desktop we will use --model=PRODUCT to identify
the firmware image to use.

Move the help text out of the "Debugging and testing options" section.
Also elaborate on the usage of remote flashing (via servo).

BUG=none
TEST=make futil
BRANCH=none

Change-Id: I817ec087104132095f8c9de2a7982f043922fbcb
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/7129820
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>