futility: Add --frid and --sku-id arguments
The purpose of identity.csv is to allow manifest key lookup in
environments without crosid (such as flashing remotely via servo). For
example, when the system firmware is completely corrupted and we want to
flash it via servo, currently there is no direct way to specify the
image to use in an archive. People would need to manually look up the
firmware manifest key (for example from ChromeOS configuration system)
and then pass `--model=MANIFEST_KEY` to futility.

To simplify the process, add --frid and --sku-id arguments. When passed,
they will override the system FRID and SKU ID obtained from the system.
Together with identity.csv, futility can use FRID and SKU ID to look up
the manifest key, and then select the correct firmware image for the
given FRID and SKU ID.

BUG=b:452553577
TEST=make futil
TEST=make runfutiltests
BRANCH=none

Change-Id: I374409637792ad9001db2c162e01f02ac623aa14
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/7129821
Reviewed-by: Jakub "Kuba" Czapiga <czapiga@google.com>