futility: Utilize identity.csv for remote device
For remote flashing, currently futility always relies on FRID matching
for model detection. That's inaccurate, because an archive may contain
multiple images with same FRID (but with different versions).

With CL:6764273, the support for identity.csv allows us to accurately
identify the model from both FRID and SKU ID. Here we extend that
feature to remote devices as well. For remote flashing, if identity.csv
exists in the archive, and both --frid and --sku-id are specified, then
futility will be able to identify the correct firmware image, even if
the existing system image is completely corrupted.

BUG=b:452553577
TEST=make futil
TEST=make runfutiltests
BRANCH=none

Change-Id: I1d92262408819e16678944b9857dd79c76122ad6
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/7129822
Reviewed-by: Jakub "Kuba" Czapiga <czapiga@google.com>